pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "superset-migrator"
version = "1.0.0"
description = "Ferramenta interativa para migrar dashboards do Apache Superset entre ambientes"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
keywords = ["superset", "dashboard", "migration", "apache-superset", "cli"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Utilities",
]
dependencies = [
    "click>=8.1",
    "httpx>=0.27",
    "pyyaml>=6.0",
    "questionary>=2.0",
    "rich>=13.0",
    "urllib3>=2.0",
]

[project.urls]
Homepage = "https://github.com/itstransdata/superset-migrator"
Repository = "https://github.com/itstransdata/superset-migrator"
"Bug Tracker" = "https://github.com/itstransdata/superset-migrator/issues"

[project.scripts]
superset-migrator = "superset_migrator.cli:run"

[tool.setuptools.packages.find]
where = ["."]
include = ["superset_migrator*"]
